+++ src/game/player.cpp
@@ -11,32 +11,32 @@ void Player::update() {
         incomingEvent(state::Event::FRAME_TIMEOUT);
     }
 
-    int x = velx_, y = vely_;
+    double x = velx_, y = vely_;
 
     if (!getStateProperties().movement_locked_x_) {
         if (Input::getInstance().isButtonHeld(input::button::LEFT)) {
             if (getStateProperties().touching_ground_) {
-                x = -10;
+                x = -10.0;
             } else {
-                x = std::max(x - 1, -10);
+                x = std::max(x - 1.0, -10.0);
             }
             incomingEvent(state::Event::MOVING);
 
             // When moving left facing_right_ should be false even when speed is zero
-            facing_right_ = x > 0;
+            facing_right_ = x > 0.0;
         } else if (Input::getInstance().isButtonHeld(input::button::RIGHT)) {
             if (getStateProperties().touching_ground_) {
-                x = 10;
+                x = 10.0;
             } else {
-                x = std::min(x + 1, 10);
+                x = std::min(x + 1.0, 10.0);
             }
             incomingEvent(state::Event::MOVING);
 
             // When moving right facing_right_ should be true even when speed is zero
-            facing_right_ = x >= 0;
+            facing_right_ = x >= 0.0;
         } else {
             if (getStateProperties().touching_ground_) {
-                x /= 5;
+                x /= 5.0;
             }
             incomingEvent(state::Event::NO_MOVEMENT);
         }
@@ -44,17 +44,17 @@ void Player::update() {
 
     if (!getStateProperties().movement_locked_y_) {
         // Gravity
-        y += 1;
+        y += 1.0;
     }
 
     if (getStateProperties().touching_wall_) {
-        y = std::min(y, 5);
+        y = std::min(y, 5.0);
     }
 
     if (Input::getInstance().isButtonPressed(input::button::DASH)) {
         if (getStateProperties().can_dash_) {
-            x = 50 * (facing_right_ ? 1 : -1);
-            y = 0;
+            x = 50.0 * (facing_right_ ? 1.0 : -1.0);
+            y = 0.0;
             incomingEvent(state::Event::DASHING);
         }
     }
@@ -62,25 +62,25 @@ void Player::update() {
     if (Input::getInstance().isButtonPressed(input::button::JUMP)) {
         if (getStateProperties().touching_wall_) {
             facing_right_ = !facing_right_;
-            int dir = facing_right_ ? 1 : -1;
-            x = 10 * dir;
-            y = -20;
+            int dir = facing_right_ ? 1.0 : -1.0;
+            x = 10.0 * dir;
+            y = -20.0;
 
             incomingEvent(state::Event::JUMPING);
         } else if (getStateProperties().can_jump_) {
-            y = -20;
+            y = -20.0;
             incomingEvent(state::Event::JUMPING);
         }
     }
 
-    y = std::max(std::min(static_cast<int>(y), 20), -20);
+    y = std::max(std::min(y, 20.0), -20.0);
     auto max_movement = getMaximumMovement(x, y, getAbsHitbox());
 
     if (max_movement.second < y) {
         incomingEvent(state::Event::TOUCHING_FLOOR);
     } else if (max_movement.first != x) {
         incomingEvent(state::Event::TOUCHING_WALL);
-    } else if (max_movement.second > 0) {
+    } else if (max_movement.second > 0.0) {
         incomingEvent(state::Event::FALLING);
     }
